refactored download image function to separate file imageDownload.js
