+++ src/front/js/utils/imageDownload.js
@@ -1,0 +1,21 @@
+export const downloadImage = async (imageURL, fileName = "transformed-image.png") => {
+    if (!imageURL) {
+      alert("No transformed image available for download.");
+      return;
+    }
+    try {
+      const response = await fetch(imageURL);
+      const blob = await response.blob();
+      const url = window.URL.createObjectURL(blob);
+      const link = document.createElement("a");
+      link.href = url;
+      link.download = fileName;
+      document.body.appendChild(link);
+      link.click();
+      document.body.removeChild(link);
+      window.URL.revokeObjectURL(url);
+    } catch (error) {
+      console.error("Error downloading the image:", error);
+      alert("Failed to download the image. Please try again.");
+    }
+  };
